Release 1.2.1 schedule select

+++ pyworkflow/project.py
@@ -419,7 +419,13 @@ class Project(object):
         2. Create the working dir and also the protocol independent db
         3. Call the launch method in protocol.job to handle submission:
            mpi, thread, queue,
-        and also take care if the execution is remotely."""
+        and also take care if the execution is remotely.
+
+        If the protocol has some prerequisited (other protocols that
+        needs to be finished first), it will be scheduled.
+        """
+        if protocol.getPrerequisites() and not scheduled:
+            return self.scheduleProtocol(protocol)
 
         isRestart = protocol.getRunMode() == MODE_RESTART
 
@@ -429,17 +435,18 @@ class Project(object):
 
         protocol.setStatus(pwprot.STATUS_LAUNCHED)
         self._setupProtocol(protocol)
-        # protocol.setMapper(self.mapper) # mapper is used in makePathAndClean
-        protocol.makePathsAndClean()  # Create working dir if necessary
-        # Delete the relations created by this protocol
-        if isRestart:
-            self.mapper.deleteRelations(self)
-        self.mapper.commit()
 
         # Prepare a separate db for this run if not from schedule jobs
         # Scheduled protocols will load the project db from the run.db file,
         # so there is no need to copy the database
+
         if not scheduled:
+            protocol.makePathsAndClean()  # Create working dir if necessary
+            # Delete the relations created by this protocol
+            if isRestart:
+                self.mapper.deleteRelations(self)
+            self.mapper.commit()
+
             # NOTE: now we are simply copying the entire project db, this can be
             # changed later to only create a subset of the db need for the run
             pwutils.path.copyFile(self.dbPath, protocol.getDbPath())
@@ -462,14 +469,16 @@ class Project(object):
             prerequisites: a list with protocols ids that the scheduled
                 protocol will wait for.
         """
+        isRestart = protocol.getRunMode() == MODE_RESTART
+
         protocol.setStatus(pwprot.STATUS_SCHEDULED)
         protocol.addPrerequisites(*prerequisites)
 
         self._setupProtocol(protocol)
-        # protocol.setMapper(self.mapper) # mapper is used in makePathAndClean
         protocol.makePathsAndClean()  # Create working dir if necessary
         # Delete the relations created by this protocol if any
-        self.mapper.deleteRelations(self)
+        if isRestart:
+            self.mapper.deleteRelations(self)
         self.mapper.commit()
 
         # Prepare a separate db for this run
@@ -969,6 +978,10 @@ class Project(object):
 
         return protocol
 
+    # FIXME: this function just return if a given object exists, not
+    # if it is a protocol, so it is incorrect judging by the name
+    # Moreover, a more consistent name (comparing to similar methods)
+    # would be: hasProtocol
     def doesProtocolExists(self, protId):
         return self.mapper.exists(protId)
 
@@ -1010,7 +1023,6 @@ class Project(object):
             protocol.addSummaryWarning(
                 "*Protocol loading problem*: A set related to this "
                 "protocol couldn't be loaded.")
-
 
     def _setupProtocol(self, protocol):
         """Insert a new protocol instance in the database"""
